@@ -15,19 +15,18 @@
     "release": "PACKAGE_VERSION=$(cat package.json | grep version | head -1 | awk -F: '{ print $2 }' | sed 's/[\",]//g' | tr -d '[[:space:]]') && git tag $PACKAGE_VERSION && git push --tags"
   },
   "dependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/plugin-transform-async-to-generator": "^7.0.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
     "animate.css": "^3.6.1",
     "api.io": "^1.0.3",
-    "async-busboy": "^0.6.2",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-async-to-module-method": "^6.16.0",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-preset-stage-1": "^6.16.0",
+    "async-busboy": "^0.7.0",
+    "babel-loader": "^8.0.2",
     "bcryptjs": "^2.4.3",
     "bluebird": "^3.4.6",
     "chart.js": "^2.7.2",
@@ -35,19 +34,18 @@
     "chron-time": "^0.0.4",
     "columnify": "^1.5.4",
     "compression-webpack-plugin": "^1.1.11",
-    "css-loader": "^0.28.11",
+    "css-loader": "^1.0.0",
     "debounce": "^1.0.2",
     "deepmerge": "^2.1.0",
     "emailjs": "^2.1.0",
     "expand-var": "^0.1.3",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^1.1.11",
+    "extract-text-webpack-plugin": "4.0.0-beta.0",
+    "file-loader": "^2.0.0",
     "file-uploader": "^1.0.0",
-    "fs-extra": "^6.0.0",
+    "fs-extra": "^7.0.0",
     "fs-extra-promise": "^1.0.1",
     "glob-promise": "^3.4.0",
     "immutability-helper": "^2.6.6",
-    "imports-loader": "^0.8.0",
     "jsx-control-statements": "^3.1.5",
     "koa": "^2.5.0",
     "koa-bodyparser": "^4.1.0",
@@ -56,7 +54,7 @@
     "koa-etag": "^3.0.0",
     "koa-range": "^0.3.0",
     "koa-route": "^3.2.0",
-    "koa-static": "^4.0.2",
+    "koa-static": "^5.0.0",
     "koa-stream": "^0.2.0",
     "koa-webpack-dev-middleware": "^2.0.1",
     "moment": "^2.22.2",
@@ -65,12 +63,12 @@
     "mousetrap": "^1.6.2",
     "postcss-cssnext": "^3.1.0",
     "postcss-each": "^0.10.0",
-    "postcss-import": "^11.1.0",
-    "postcss-loader": "^2.1.3",
+    "postcss-import": "^12.0.0",
+    "postcss-loader": "^3.0.0",
     "postcss-mixins": "^6.2.0",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
-    "react-dropzone": "^4.2.11",
+    "react-dropzone": "^5.0.1",
     "react-google-maps": "^9.4.5",
     "react-notification": "^6.8.2",
     "react-region-select": "^2.2.0",
@@ -80,29 +78,29 @@
     "request": "^2.85.0",
     "request-promise-native": "^1.0.4",
     "semantic-ui-css": "^2.3.1",
-    "semantic-ui-react": "^0.81.1",
+    "semantic-ui-react": "^0.82.3",
     "server-destroy": "^1.0.1",
-    "sftp-fs": "^0.0.5",
+    "sftp-fs": "^0.0.6",
     "sha1": "^1.1.1",
     "shuffle-array": "^1.0.1",
     "simple-xml2json": "^1.2.3",
     "source-map-support": "^0.5.4",
     "store": "^2.0.12",
     "strip-ansi": "^4.0.0",
-    "style-loader": "^0.21.0",
+    "style-loader": "^0.23.0",
     "superagent": "^3.8.2",
     "uglify-js": "^3.4.0",
     "url-loader": "^1.0.1",
     "uuid": "^3.2.1",
     "walk-promise": "^0.2.0",
-    "webpack": "^4.12.0",
+    "webpack": "^4.17.2",
     "wsh.js": "^0.0.8",
-    "yargs": "^11.0.0"
+    "yargs": "^12.0.1"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.2",
+    "babel-eslint": "^9.0.0",
     "duplicate-package-checker-webpack-plugin": "^3.0.0",
-    "eslint": "^4.19.1",
+    "eslint": "^5.5.0",
     "eslint-plugin-babel": "^5.0.0",
     "eslint-plugin-jsx-control-statements": "^2.2.0",
     "eslint-plugin-react": "^7.9.1",
