@@ -12,7 +12,7 @@ class Database {
     async init(config) {
         const url = config.mongoUrl;
 
-        this.client = await MongoClient.connect(url);
+        this.client = await MongoClient.connect(url, { useNewUrlParser: true });
         this.db = this.client.db(path.basename(url));
     }
 
